requirements.txt
vyper==0.4.0
eth-ape==0.8.24
titanoboa==0.2.5
snekmate==0.1.0
pytest==8.3.4
